% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm-methods.R
\name{summary.gm_fit}
\alias{summary.gm_fit}
\alias{print.gm_fit}
\alias{logLik.gm_fit}
\alias{coef.gm_fit}
\alias{vcov.gm_fit}
\title{Summary of a GARCH-MIDAS fit}
\usage{
\method{summary}{gm_fit}(object, ...)

\method{print}{gm_fit}(x, ...)

\method{logLik}{gm_fit}(object, ...)

\method{coef}{gm_fit}(object, ...)

\method{vcov}{gm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{gm_fit} object.}

\item{...}{Unused.}

\item{x}{A \code{gm_fit} object.}
}
\value{
\code{object}, invisibly.
}
\description{
Prints the robust coefficient table with significance codes, sample
information and in-sample losses, in the same layout as the legacy
\code{summary.gmnoskew()}.
}
\section{Functions}{
\itemize{
\item \code{print(gm_fit)}: Compact coefficient printout.

\item \code{logLik(gm_fit)}: Log-likelihood with df/nobs attributes.

\item \code{coef(gm_fit)}: Estimated coefficients.

\item \code{vcov(gm_fit)}: Robust (sandwich) variance-covariance matrix.

}}
