% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topix_autos}
\alias{topix_autos}
\title{Daily TOPIX-17 Autos & Transportation Equipment returns}
\format{
An \code{xts} object with one column \code{AUTOS}.
}
\source{
Tokyo Stock Exchange, via Datastream.
}
\usage{
topix_autos
}
\description{
Daily returns of the TOPIX-17 "Autos & Transportation Equipment" industry
index, used as the example industry asset.
日频 TOPIX-17 汽车与运输设备行业指数收益,示例中作为行业资产。
}
\seealso{
\link{topix_market}
}
\keyword{datasets}
